changed FS_OpenFile() to return long instead of size_t,
and adjusted fs_filesize, and other places accordingly.

+++ engine/h2shared/midi_camd.c
@@ -786,7 +786,7 @@ static void *MIDI_Play (const char *filename)
 
 	struct SMFHeader *hdr;
 	UBYTE *pbyte;
-	size_t smfdatasize;
+	long smfdatasize;
 	UWORD track;
 	LONG error;
 	ULONG id;
